requirements.txt
flask>=3.1.3
openai>=1.75.0
playwright>=1.58.0
python-dotenv>=1.0.1
